files over 2gb should be supported now. receive and reconstruct still needs to be implemented.

+++ RsyncNet/Delta/CopyDelta.cs
@@ -4,7 +4,7 @@
     {
         #region IDelta Properties, indexers, events and operators
 
-        public long Length { get; set; }
+        public int Length { get; set; }
         public long Offset { get; set; }
 
         #endregion
